finished the project
finished the project for the most part, just need to do the paperwork side of things but thats unimportant for this.

major changes include removing PopArray seeing as I had no actual use for it
completing setup.java and checking that all of the methods of Population are actually usable.